+++ src/commonMain/kotlin/exchange.dydx.abacus/state/StateManagerAdaptorV2.kt
@@ -1,7 +1,6 @@
 package exchange.dydx.abacus.state
 
 import exchange.dydx.abacus.output.Compliance
-import exchange.dydx.abacus.output.ComplianceAction
 import exchange.dydx.abacus.output.ComplianceStatus
 import exchange.dydx.abacus.output.Notification
 import exchange.dydx.abacus.output.PerpetualState
@@ -84,7 +83,6 @@ import exchange.dydx.abacus.state.supervisor.stopWatchingLastOrder
 import exchange.dydx.abacus.state.supervisor.subaccountNumber
 import exchange.dydx.abacus.state.supervisor.subaccountTransferPayload
 import exchange.dydx.abacus.state.supervisor.trade
-import exchange.dydx.abacus.state.supervisor.triggerCompliance
 import exchange.dydx.abacus.state.supervisor.triggerOrders
 import exchange.dydx.abacus.state.supervisor.triggerOrdersPayload
 import exchange.dydx.abacus.state.supervisor.walletConnectionType
@@ -678,10 +676,6 @@ internal class StateManagerAdaptorV2(
 
     internal fun screen(address: String, callback: (restriction: Restriction) -> Unit) {
         accounts.screen(address, callback)
-    }
-
-    internal fun triggerCompliance(action: ComplianceAction, callback: TransactionCallback?) {
-        accounts.triggerCompliance(action, callback)
     }
 
     internal fun registerPushNotification(token: String, languageCode: String?) {
